# HIVE-28548: Subdivide memory size allocated to parallel operators