@okumin 2025-08-11
@ngsg @deniskuzZ @abstractdog Do you think this approach is reasonable? If the answer is yes, I will rebase this branch