@ngsg 2025-08-12
I think this is a reasonable approach.

Regarding the `isTez`/`isLlap` issue in `VectorGroupByOperator`, I'm inclined to use `isTez` and drop `isLlap`, as `VectorGroupByOperator` checks overall heap usage via `gcCanary`, not `MemoryMXBean#getHeapMemoryUsage` and `memoryThreshold`. However, I'm not certain whether Tez without LLAP also respects `getConf().getMaxMemoryAvailable()`, so it would be nice if others could confirm this as well.